Initial vite setup and file moving
