+++ src/components/AlgoliaLogo.js
