+++ src/components/FormidableLogo.js
@@ -43,7 +43,17 @@ export default html`
             L197.21,199.31z"
       ></path>
     </g>
-    <circle cx="223.3" cy="26" r="5.5" fill="none" stroke="rgba(255,255,255,0.1)" stroke-width="2"/>
-    <path fill="rgba(255,255,255,0.1)" d="M223.1,25.6c0.3,0,0.5-0.1,0.6-0.2c0.1-0.1,0.2-0.3,0.2-0.5c0-0.2-0.1-0.4-0.2-0.5c-0.1-0.1-0.3-0.1-0.6-0.1       h-0.6v1.3H223.1 M222.5,26.5v2h-1.3v-5.2h2c0.7,0,1.2,0.1,1.5,0.3c0.3,0.2,0.5,0.6,0.5,1.1c0,0.3-0.1,0.6-0.2,0.8       c-0.2,0.2-0.4,0.4-0.7,0.5c0.2,0,0.3,0.1,0.5,0.3c0.1,0.1,0.3,0.4,0.4,0.7l0.7,1.5h-1.4l-0.6-1.3c-0.1-0.3-0.3-0.4-0.4-0.5       c-0.1-0.1-0.3-0.1-0.5-0.1H222.5"/>
+    <circle
+      cx="223.3"
+      cy="26"
+      r="5.5"
+      fill="none"
+      stroke="rgba(255,255,255,0.1)"
+      stroke-width="2"
+    />
+    <path
+      fill="rgba(255,255,255,0.1)"
+      d="M223.1,25.6c0.3,0,0.5-0.1,0.6-0.2c0.1-0.1,0.2-0.3,0.2-0.5c0-0.2-0.1-0.4-0.2-0.5c-0.1-0.1-0.3-0.1-0.6-0.1       h-0.6v1.3H223.1 M222.5,26.5v2h-1.3v-5.2h2c0.7,0,1.2,0.1,1.5,0.3c0.3,0.2,0.5,0.6,0.5,1.1c0,0.3-0.1,0.6-0.2,0.8       c-0.2,0.2-0.4,0.4-0.7,0.5c0.2,0,0.3,0.1,0.5,0.3c0.1,0.1,0.3,0.4,0.4,0.7l0.7,1.5h-1.4l-0.6-1.3c-0.1-0.3-0.3-0.4-0.4-0.5       c-0.1-0.1-0.3-0.1-0.5-0.1H222.5"
+    />
   </svg>
 `;
