+++ src/components/SearchIcon.js
