+++ src/utils/analyzeFile.js
